Changed to Int64 in CoreData
For saving larger tire mileage values. Also changed Picker in EditTireView so .inStorage and .onVehicle orientation matched what is in AddTireView. They were switched.

+++ Tyre Myles/DataModel/TireModel.swift
@@ -63,9 +63,9 @@ final class DataModel: NSObject, ObservableObject, NSFetchedResultsControllerDel
 		let entity = TireEntity(context: moc)
 
 		entity.name = name
-		// CoreData model defines these values as Int16
-		entity.installMiles = Int16(installMiles) ?? 0
-		entity.removalMiles = Int16(removalMiles) ?? 0
+		// CoreData model defines these values as Int64
+		entity.installMiles = Int64(installMiles) ?? 0
+		entity.removalMiles = Int64(removalMiles) ?? 0
 		entity.seasonType = season.rawValue
 		entity.installDate = installDate
 		entity.removalDate = removallDate
